Add mocks package
